instacommerce_ai/backend/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
bcrypt==4.1.3
certifi==2025.1.31
charset-normalizer==3.4.1
distro==1.9.0
fastapi==0.110.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.1.17
langchain-core==0.1.49
langchain-openai==0.0.8
langsmith==0.1.66
openai==1.65.1
orjson==3.10.15
packaging==23.2
pinecone-client==3.2.2
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
sniffio==1.3.1
tenacity==8.2.3
tiktoken==0.9.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.29.0
zstandard==0.23.0
passlib[bcrypt]==1.7.4
